+++ src/NewPicker/PickerInput/hooks/useRangeValue.ts
@@ -2,7 +2,7 @@ import { useEvent, useMergedState } from 'rc-util';
 import * as React from 'react';
 import type { GenerateConfig } from '../../../generate';
 import { formatValue, isSame, isSameTimestamp } from '../../../utils/dateUtil';
-import useSyncState from '../../hooks/useSyncState';
+import useSyncState from '../../../hooks/useSyncState';
 import type { BaseInfo, FormatType, Locale, ReplaceListType } from '../../../interface';
 import { fillIndex } from '../../util';
 import type { RangePickerProps } from '../RangePicker';
