+++ src/NewPicker/PickerPanel/index.tsx
@@ -14,7 +14,7 @@ import type {
 } from '../../interface';
 import { isSame } from '../../utils/dateUtil';
 import { toArray } from '../../utils/miscUtil';
-import useLocale from '../hooks/useLocale';
+import useLocale from '../../hooks/useLocale';
 import { getTimeConfig } from '../hooks/useTimeConfig';
 import useToggleDates from '../hooks/useToggleDates';
 import PickerContext from '../PickerInput/context';
